move _union_categoricals to cudf.api.types
