+++ python/cudf/cudf/api/__init__.py
@@ -1,1 +1,3 @@
 # Copyright (c) 2020, NVIDIA CORPORATION.
+
+from cudf.api import types
