+++ python/cudf/cudf/api/types/__init__.py
@@ -1,0 +1,3 @@
+# Copyright (c) 2021, NVIDIA CORPORATION.
+
+from .categoricals import _union_categoricals
